# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Waste Collection API",
        "description": "API Documentation for E-Waste & Scrap Collection Platform",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/donation-request": {
            "post": {
                "tags": [
                    "Donation"
                ],
                "summary": "Create a new donation pickup request",
                "operationId": "createDonationRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "address",
                                    "city_id",
                                    "scheduled_at",
                                    "donation_category",
                                    "items",
                                    "images"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, City"
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 19.076
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": 72.8777
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-20 10:00:00"
                                    },
                                    "donation_category": {
                                        "type": "string",
                                        "example": "clothes"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "description": "Min 1, max 3 images (max 5MB each)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 21
                                                },
                                                "weight": {
                                                    "type": "number",
                                                    "example": 10.5
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation Request Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/donation-requests": {
            "get": {
                "tags": [
                    "Donation"
                ],
                "summary": "List donation pickup requests",
                "operationId": "getDonationRequests",
                "responses": {
                    "200": {
                        "description": "List of donations"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/pickup-requests/{id}/clone-as-donation": {
            "post": {
                "tags": [
                    "Donation"
                ],
                "summary": "Clone an existing scrap pickup as a donation",
                "operationId": "cloneAsDonation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation Created"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}